Github/Voice Based AI agent/requirements.txt
openai>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2
numpy>=1.21.0
scikit-learn>=0.24.2
python-dotenv>=0.19.0
openai-whisper>=20231117
torch>=2.0.0
langdetect>=1.0.9
gTTS>=2.3.2
pydub>=0.25.1 